backend/requirements.txt
numpy==1.26.4
python-dotenv==1.0.1
langchain==0.2.0
langchain-openai==0.1.7
langchain-community==0.2.0
langgraph==0.2.3
supabase==2.4.0
pypdf==4.2.0